refactor(scripts): extract sanitizePromptInput into a shared module
The prompt-input sanitizer was duplicated in bedrock_classifier.ts and
detect_duplicates.ts; the two copies had already drifted by a stray comment.
Move it into a single sanitize_utils.ts and import it from both call sites so
there is one implementation to maintain and review.

No behaviour change: the extracted function body is identical to the previous
copies (verified by `npm run build` and the existing jest suite).
