+++ scripts/bedrock_classifier.ts
@@ -9,6 +9,7 @@ import {
 } from "@aws-sdk/client-bedrock-runtime";
 import { ClassificationResult, LabelTaxonomy } from "./data_models.js";
 import { retryWithBackoff } from "./retry_utils.js";
+import { sanitizePromptInput } from "./sanitize_utils.js";
 
 const MODEL_ID = "us.anthropic.claude-sonnet-4-20250514-v1:0";
 const MAX_TOKENS = 2048;
@@ -18,49 +19,6 @@ const TOP_P = 0.9;
 // Security: Maximum lengths for input validation
 const MAX_TITLE_LENGTH = 500;
 const MAX_BODY_LENGTH = 10000;
-
-/**
- * Sanitize user input to prevent prompt injection attacks
- */
-function sanitizePromptInput(input: string, maxLength: number): string {
-  if (!input) {
-    return "";
-  }
-
-  // Truncate to maximum length
-  let sanitized = input.substring(0, maxLength);
-
-  // Remove potential prompt injection patterns
-  const dangerousPatterns = [
-    /ignore\s+(all\s+)?(previous|above|prior)\s+instructions?/gi,
-    /disregard\s+(all\s+)?(previous|above|prior)\s+instructions?/gi,
-    /forget\s+(all\s+)?(previous|above|prior)\s+instructions?/gi,
-    /new\s+instructions?:/gi,
-    /system\s*:/gi,
-    /assistant\s*:/gi,
-    /\[SYSTEM\]/gi,
-    /\[ASSISTANT\]/gi,
-    /\<\|im_start\|\>/gi,
-    /\<\|im_end\|\>/gi,
-  ];
-
-  for (const pattern of dangerousPatterns) {
-    sanitized = sanitized.replace(pattern, "[REDACTED]");
-  }
-
-  // Escape backticks that could break JSON formatting
-  sanitized = sanitized.replace(/`/g, "'");
-
-  // Remove excessive newlines that could break prompt structure
-  sanitized = sanitized.replace(/\n{4,}/g, "\n\n\n");
-
-  // Add truncation notice if content was cut
-  if (input.length > maxLength) {
-    sanitized += "\n\n[Content truncated for security]";
-  }
-
-  return sanitized;
-}
 
 /**
  * Initialize Bedrock client with AWS credentials
